+++ src/app/components/quotation/TopBar.tsx
@@ -1,0 +1,7 @@
+export default function TopBar() {
+  return (
+    <div className="fixed h-8 right-4 py-3 px-6 flex flex-col">
+      <p>(주)현중푸드/성남시</p>
+    </div>
+  );
+}
